references/sample_code_for_text_extracting/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
filelock==3.3.1
gensim==4.1.2
hanlp-downloader==0.0.25
hanziconv==0.3.2
harvesttext==0.8.1.4
huggingface-hub==0.0.19
idna==3.3
importlib-metadata==4.8.1
isodate==0.6.0
jieba==0.42.1
joblib==1.1.0
JPype1==0.7.0
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
nltk==3.6.5
numpy==1.21.3
opencc-python-reimplemented==0.1.6
packaging==21.0
pandas==1.1.5
Pillow==8.4.0
pyhanlp==0.1.79
pyparsing==3.0.3
pypinyin==0.43.0
python-dateutil==2.8.2
python-louvain==0.15
pytz==2021.3
PyYAML==6.0
rdflib==6.0.2
regex==2021.10.23
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0
tqdm==4.62.3
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
w3lib==1.22.0
wincertstore==0.2
zipp==3.6.0
